backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pyjwt[crypto]==2.9.0
bcrypt==4.2.0
redis==5.1.1
supabase==2.9.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
